pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "pandas>=2.2.3",
    "paramiko>=3.5.1",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
]
